# shared/generated/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Title",
    "description": "Title",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3333/api"
    }
  ],
  "paths": {
...
    "/user/me/getLinkingCode": {
      "get": {
        "description": "get the current user linking code",
        "operationId": "getLinkingCode",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid user supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/crypto/cryptos": {
      "put": {
        "tags": [
          "crypto"
        ],
        "operationId": "setCryptos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CryptoHolding"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CryptoHolding"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Crypto not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/crypto/message": {
      "put": {
        "tags": [
          "crypto"
        ],
        "operationId": "setMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Crypto not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/crypto/{id}/value": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Crypto ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "crypto"
        ],
        "operationId": "getCoinValue",
        "responses": {
          "200": {
            "description": "A successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Crypto not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "CryptoHolding": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "token": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "price": {
            "type": "number"
          }
        }
      },
...
    }
  }
}